Home HTML & CSS
Still working on it but I've made some  changes according to Tommi's requests. So we have to add a Navigation Bar to the page. Due to that I had to mage the overall Body Width to 1920px. So please follow that body size on your pages as well. I'm adding the Home page HTML and CSS here. You can get the necessary codes from it if you need. Right now the Nav bar is on top of the page. But I'll take it down and align to the Home and make reservations buttons. Let me know if you have any questions.